requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
black==25.1.0
cffi==1.16.0
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
constantly==23.10.4
cryptography==41.0.7
daphne==3.0.2
Django==4.2.7
django-axes==7.0.1
django-environ==0.12.0
gunicorn==22.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.1.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
pyotp==2.9.0
pypng==0.20220715.0
PyQRCode==1.2.1
qrcode==7.4.2
redis==5.2.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.34.0
zope.interface==6.1
